removed rdrs_date.hpp, now using my_time.h
